@@ -30,101 +31,114 @@
     "clean": "rimraf ./dist && rimraf ./node_modules && npm cache clean --force && git checkout -- package.json && git checkout -- package-lock.json",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s -r 0",
     "changelog-current": "conventional-changelog -p angular -r 2",
+    "preversion.linux": "npm run test.all && rimraf ./dist && npm run build.prod",
+    "version.linux": "npm run changelog && git add -A CHANGELOG.md",
+    "postversion.linux": "git push && git push --tags && rimraf igo2.zip && cd ./dist/igo2/ && zip -r ../igo2.zip ./* && npm run ghpages",
     "preversion": "npm run test.all && rimraf ./dist && npm run build.prod",
     "version": "npm run changelog && git add -A CHANGELOG.md",
-    "postversion": "git push && git push --tags && rimraf igo2.zip && cd ./dist/igo2/ && zip -r ../igo2.zip ./* && npm run ghpages && npm run ghrelease",
+    "postversion": "git push && git push --tags && rimraf igo2.zip && cd ./dist/igo2/ && tar -a -c -f ../igo2.zip * && npm run ghpages.win",
     "preghpages": "npm run build -- --configuration=github --output-path ./dist/ghpages --base-href /igo2/ && echo \"include: ['_default.json', '_contexts.json', '_base.json']\" > dist/ghpages/_config.yml",
     "ghpages": "VERSION=$(node -p -e \"require('./package.json').version\") && npx ngh --dir=dist/ghpages --no-silent=false --message=$VERSION",
     "postghpages": "rimraf ./dist/ghpages",
-    "ghrelease": "VERSION=$(node -p -e \"require('./package.json').version\") && publish-release --owner infra-geo-ouverte --repo igo2 --tag $VERSION --name $VERSION --reuse-release --notes \"$(npm run changelog-current -s)\" --assets dist/igo2.zip"
+    "preghpages.win": "npm run build -- --configuration=github --output-path ./dist/ghpages --base-href /igo2/ && echo include: ['_default.json', '_contexts.json', '_base.json'] > dist/ghpages/_config.yml",
+    "ghpages.win": "npx ngh --dir=dist/ghpages --no-silent=false --message=%npm_package_version%",
+    "postghpages.win": "rimraf ./dist/ghpages"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.3",
-    "@angular/cdk": "^10.0.1",
-    "@angular/common": "^10.0.3",
-    "@angular/compiler": "^10.0.3",
-    "@angular/core": "^10.0.3",
-    "@angular/forms": "^10.0.3",
-    "@angular/material": "^10.0.1",
-    "@angular/platform-browser": "^10.0.3",
-    "@angular/platform-browser-dynamic": "^10.0.3",
-    "@angular/router": "^10.0.3",
-    "@angular/service-worker": "^10.0.3",
-    "@azure/msal-angular": "^2.0.0",
-    "@azure/msal-browser": "^2.14.2",
-    "@igo2/auth": "~1.8.2",
-    "@igo2/common": "~1.8.2",
-    "@igo2/context": "~1.8.2",
-    "@igo2/core": "~1.8.2",
-    "@igo2/geo": "~1.8.2",
-    "@igo2/integration": "~1.8.2",
-    "@igo2/utils": "~1.8.2",
-    "@mat-datetimepicker/core": "^5.1.2",
-    "@turf/buffer": "^5.1.5",
-    "@turf/point-on-feature": "^5.1.5",
-    "angular-shepherd": "^0.7.0",
-    "bowser": "^2.9.0",
+    "@angular/animations": "^12.2.9",
+    "@angular/cdk": "^12.2.9",
+    "@angular/common": "^12.2.9",
+    "@angular/compiler": "^12.2.9",
+    "@angular/core": "^12.2.9",
+    "@angular/forms": "^12.2.9",
+    "@angular/material": "^12.2.9",
+    "@angular/platform-browser": "^12.2.9",
+    "@angular/platform-browser-dynamic": "^12.2.9",
+    "@angular/router": "^12.2.9",
+    "@angular/service-worker": "^12.2.9",
+    "@azure/msal-angular": "^2.0.1",
+    "@azure/msal-browser": "^2.15.0",
+    "@igo2/auth": "~1.9.3",
+    "@igo2/common": "~1.9.3",
+    "@igo2/context": "~1.9.3",
+    "@igo2/core": "~1.9.3",
+    "@igo2/geo": "~1.9.3",
+    "@igo2/integration": "~1.9.3",
+    "@igo2/utils": "~1.9.3",
+    "@mat-datetimepicker/core":"^7.0.1",
+    "@turf/buffer": "^6.5.0",
+    "@turf/point-on-feature": "^6.5.0",
+    "angular-shepherd": "12.0.0",
+    "bowser": "^2.10.0",
     "classlist.js": "^1.1.20150312",
     "core-js": "^3.6.5",
     "fast-text-encoding": "^1.0.3",
     "file-saver": "^2.0.2",
+    "gulp-concat": "^2.6.1",
+    "gulp-replace": "^1.1.3",
     "hammerjs": "^2.0.8",
     "html2canvas": "^1.0.0-rc.5",
-    "http-server": "^0.12.3",
-    "jspdf": "^1.5.3",
+    "http-server": "^13.0.2",
+    "jspdf": "^2.4.0",
     "jszip": "^3.1.5",
     "jwt-decode": "^2.2.0",
     "moment": "^2.24.0",
     "natives": "^1.1.6",
-    "ngx-cacheable": "^1.0.9",
     "ngx-color-picker": "^10.1.0",
-    "ol": "^6.3.1",
-    "ol-mapbox-style": "^6.1.2",
-    "proj4": "^2.6.2",
-    "rxjs": "^6.6.0",
+    "ol": "^6.8.1",
+    "ol-mapbox-style": "^6.5.1",
+    "proj4": "^2.7.5",
+    "rxjs": "^6.6.7",
     "scroll-into-view-if-needed": "^2.2.20",
     "stream": "^0.0.2",
+    "striptags": "^3.2.0",
+    "ts-cacheable": "^1.0.5",
     "tslib": "^2.0.0",
     "typy": "^3.3.0",
     "unorm": "^1.4.1",
     "web-animations-js": "^2.3.2",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.1",
-    "@angular-devkit/build-ng-packagr": "~0.1000.1",
-    "@angular/cli": "^10.0.1",
-    "@angular/compiler-cli": "^10.0.3",
+    "@angular-devkit/build-angular": "~12.2.9",
+    "@angular-eslint/builder": "12.5.0",
+    "@angular-eslint/eslint-plugin": "12.5.0",
+    "@angular-eslint/eslint-plugin-template": "12.5.0",
+    "@angular-eslint/schematics": "12.5.0",
+    "@angular-eslint/template-parser": "12.5.0",
+    "@angular/cli": "^12.2.9",
+    "@angular/compiler-cli": "^12.2.9",
     "@babel/core": "^7.7.4",
     "@babel/preset-env": "^7.7.4",
+    "@compodoc/compodoc": "^1.1.15",
     "@types/geojson": "^7946.0.6",
     "@types/hammerjs": "^2.0.36",
     "@types/jasmine": "^3.5.11",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^14.0.22",
-    "angular-cli-ghpages": "^0.6.2",
-    "codelyzer": "^6.0.0",
-    "compodoc": "0.0.41",
+    "@typescript-eslint/eslint-plugin": "4.28.2",
+    "@typescript-eslint/parser": "4.28.2",
+    "angular-cli-ghpages": "^1.0.0-rc.2",
     "conventional-changelog-cli": "^2.0.12",
-    "gulp": "^4.0.2",
+    "eslint": "^7.26.0",
+    "eslint-plugin-unused-imports": "^1.1.5",
+    "gulp": "^4.0.0",
     "gulp-babel": "^8.0.0",
-    "gulp-json-editor": "^2.4.2",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "^5.1.0",
+    "karma": "^6.3.4",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^10.0.0",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "ng-packagr": "^12.1.2",
     "npx": "^10.2.0",
     "protractor": "~7.0.0",
-    "publish-release": "^1.6.0",
-    "ts-node": "^8.10.2",
-    "tslint": "~6.1.0",
-    "typescript": "~3.9.6",
-    "webdriver-manager": "^12.1.7",
-    "webpack": "^4.41.2"
+    "sass": "^1.42.1",
+    "ts-node": "^10.1.0",
+    "typescript": "~4.3.5",
+    "webdriver-manager": "^12.1.6",
+    "webpack": "^5.58.1"
   }
 }
