@@ -10,7 +10,8 @@ import {
   IgoPanelModule,
   IgoFlexibleModule,
   IgoToolModule,
-  IgoInteractiveTourModule
+  IgoInteractiveTourModule,
+  IgoHomeButtonModule
 } from '@igo2/common';
 import { IgoFeatureModule } from '@igo2/geo';
 import { IgoContextManagerModule } from '@igo2/context';
@@ -30,7 +31,8 @@ import { SidenavComponent } from './sidenav.component';
     IgoContextManagerModule,
     IgoToolModule,
     IgoFeatureModule,
-    IgoInteractiveTourModule
+    IgoInteractiveTourModule,
+    IgoHomeButtonModule
   ],
   exports: [SidenavComponent],
   declarations: [SidenavComponent]
