@@ -404,7 +404,7 @@ export class ToastPanelComponent implements OnInit, OnDestroy {
         ),
         icon: 'arrow-expand',
         display: () => {
-          return this.fullExtent$.pipe(map((v) => !v  && !this.isDesktop()));
+          return this.fullExtent$.pipe(map((v) => !v && !this.isDesktop()));
         },
         handler: () => {
           this.fullExtent = true;
@@ -451,7 +451,7 @@ export class ToastPanelComponent implements OnInit, OnDestroy {
     }
     const myOlFeature = featureToOl(result.data, this.map.projection);
     const olGeometry = myOlFeature.getGeometry();
-    if (featuresAreTooDeepInView(this.map, olGeometry.getExtent(), 0.0025)) {
+    if (featuresAreTooDeepInView(this.map, olGeometry.getExtent() as [number, number, number, number], 0.0025)) {
       const extent = olGeometry.getExtent();
       const x = extent[0] + (extent[2] - extent[0]) / 2;
       const y = extent[1] + (extent[3] - extent[1]) / 2;
